+++ op-node/sources/l2_client.go
@@ -42,8 +42,8 @@ func L2ClientDefaultConfig(config *rollup.Config, trustRPC bool) *L2ClientConfig
 	return &L2ClientConfig{
 		EthClientConfig: EthClientConfig{
 			// receipts and transactions are cached per block
-			ReceiptsCacheVolumeByte:     span * 1024 * 1024,
-			TransactionsCacheVolumeByte: span * 1024 * 1024,
+			ReceiptsCacheVolumeByte:     uint64(span * 1024 * 1024),
+			TransactionsCacheVolumeByte: uint64(span * 1024 * 1024),
 			HeadersCacheSize:            span,
 			PayloadsCacheSize:           span,
 			MaxRequestsPerBatch:         20, // TODO: tune batch param
